+++ src/terminal.go
@@ -24,7 +24,7 @@ import (
 var placeholder *regexp.Regexp
 
 func init() {
-	placeholder = regexp.MustCompile("\\\\?(?:{\\+?[0-9,-.]*}|{q})")
+	placeholder = regexp.MustCompile("\\\\?(?:{[+s]*[0-9,-.]*}|{q})")
 }
 
 type jumpMode int
@@ -59,7 +59,8 @@ type Terminal struct {
 	inlineInfo bool
 	prompt     string
 	promptLen  int
-	reverse    bool
+	queryLen   [2]int
+	layout     layoutType
 	fullscreen bool
 	hscroll    bool
 	hscrollOff int
@@ -68,6 +69,7 @@ type Terminal struct {
 	cx         int
 	cy         int
 	offset     int
+	xoffset    int
 	yanked     []rune
 	input      []rune
 	multi      bool
@@ -221,6 +223,12 @@ const (
 	actTop
 )
 
+type placeholderFlags struct {
+	plus          bool
+	preserveSpace bool
+	query         bool
+}
+
 func toActions(types ...actionType) []action {
 	actions := make([]action, len(types))
 	for idx, t := range types {
@@ -277,6 +285,9 @@ func defaultKeymap() map[int][]action {
 	keymap[tui.PgUp] = toActions(actPageUp)
 	keymap[tui.PgDn] = toActions(actPageDown)
 
+	keymap[tui.SUp] = toActions(actPreviewUp)
+	keymap[tui.SDown] = toActions(actPreviewDown)
+
 	keymap[tui.Rune] = toActions(actRune)
 	keymap[tui.Mouse] = toActions(actMouse)
 	keymap[tui.DoubleClick] = toActions(actAccept)
@@ -293,10 +304,11 @@ func trimQuery(query string) []rune {
 func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	input := trimQuery(opts.Query)
 	var header []string
-	if opts.Reverse {
-		header = opts.Header
-	} else {
+	switch opts.Layout {
+	case layoutDefault, layoutReverseList:
 		header = reverseStringArray(opts.Header)
+	default:
+		header = opts.Header
 	}
 	var delay time.Duration
 	if opts.Tac {
@@ -354,7 +366,8 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	t := Terminal{
 		initDelay:  delay,
 		inlineInfo: opts.InlineInfo,
-		reverse:    opts.Reverse,
+		queryLen:   [2]int{0, 0},
+		layout:     opts.Layout,
 		fullscreen: fullscreen,
 		hscroll:    opts.Hscroll,
 		hscrollOff: opts.HscrollOff,
@@ -363,6 +376,7 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		cx:         len(input),
 		cy:         0,
 		offset:     0,
+		xoffset:    0,
 		yanked:     []rune{},
 		input:      input,
 		multi:      opts.Multi,
@@ -630,12 +644,24 @@ func (t *Terminal) resizeWindows() {
 	for i := 0; i < t.window.Height(); i++ {
 		t.window.MoveAndClear(i, 0)
 	}
-	t.truncateQuery()
 }
 
 func (t *Terminal) move(y int, x int, clear bool) {
-	if !t.reverse {
-		y = t.window.Height() - y - 1
+	h := t.window.Height()
+
+	switch t.layout {
+	case layoutDefault:
+		y = h - y - 1
+	case layoutReverseList:
+		n := 2 + len(t.header)
+		if t.inlineInfo {
+			n--
+		}
+		if y < n {
+			y = h - y - 1
+		} else {
+			y -= n
+		}
 	}
 
 	if clear {
@@ -645,20 +671,44 @@ func (t *Terminal) move(y int, x int, clear bool) {
 	}
 }
 
+func (t *Terminal) truncateQuery() {
+	t.input, _ = t.trimRight(t.input, maxPatternLength)
+	t.cx = util.Constrain(t.cx, 0, len(t.input))
+}
+
+func (t *Terminal) updatePromptOffset() ([]rune, []rune) {
+	maxWidth := util.Max(1, t.window.Width()-t.promptLen-1)
+
+	_, overflow := t.trimLeft(t.input[:t.cx], maxWidth)
+	minOffset := int(overflow)
+	maxOffset := util.Min(util.Min(len(t.input), minOffset+maxWidth), t.cx)
+
+	t.xoffset = util.Constrain(t.xoffset, minOffset, maxOffset)
+	before, _ := t.trimLeft(t.input[t.xoffset:t.cx], maxWidth)
+	beforeLen := t.displayWidth(before)
+	after, _ := t.trimRight(t.input[t.cx:], maxWidth-beforeLen)
+	afterLen := t.displayWidth(after)
+	t.queryLen = [2]int{beforeLen, afterLen}
+	return before, after
+}
+
 func (t *Terminal) placeCursor() {
-	t.move(0, t.promptLen+t.displayWidth(t.input[:t.cx]), false)
+	t.move(0, t.promptLen+t.queryLen[0], false)
 }
 
 func (t *Terminal) printPrompt() {
 	t.move(0, 0, true)
 	t.window.CPrint(tui.ColPrompt, t.strong, t.prompt)
-	t.window.CPrint(tui.ColNormal, t.strong, string(t.input))
+
+	before, after := t.updatePromptOffset()
+	t.window.CPrint(tui.ColNormal, t.strong, string(before))
+	t.window.CPrint(tui.ColNormal, t.strong, string(after))
 }
 
 func (t *Terminal) printInfo() {
 	pos := 0
 	if t.inlineInfo {
-		pos = t.promptLen + t.displayWidth(t.input) + 1
+		pos = t.promptLen + t.queryLen[0] + t.queryLen[1] + 1
 		if pos+len(" < ") > t.window.Width() {
 			return
 		}
@@ -739,7 +789,7 @@ func (t *Terminal) printList() {
 	count := t.merger.Length() - t.offset
 	for j := 0; j < maxy; j++ {
 		i := j
-		if !t.reverse {
+		if t.layout == layoutDefault {
 			i = maxy - 1 - j
 		}
 		line := i + 2 + len(t.header)
@@ -1127,16 +1177,46 @@ func quoteEntry(entry string) string {
 	return "'" + strings.Replace(entry, "'", "'\\''", -1) + "'"
 }
 
-func hasPlusFlag(template string) bool {
-	for _, match := range placeholder.FindAllString(template, -1) {
-		if match[0] == '\\' {
-			continue
-		}
-		if match[1] == '+' {
-			return true
+func parsePlaceholder(match string) (bool, string, placeholderFlags) {
+	flags := placeholderFlags{}
+
+	if match[0] == '\\' {
+		// Escaped placeholder pattern
+		return true, match[1:], flags
+	}
+
+	skipChars := 1
+	for _, char := range match[1:] {
+		switch char {
+		case '+':
+			flags.plus = true
+			skipChars++
+		case 's':
+			flags.preserveSpace = true
+			skipChars++
+		case 'q':
+			flags.query = true
+		default:
+			break
 		}
 	}
-	return false
+
+	matchWithoutFlags := "{" + match[skipChars:]
+
+	return false, matchWithoutFlags, flags
+}
+
+func hasPreviewFlags(template string) (plus bool, query bool) {
+	for _, match := range placeholder.FindAllString(template, -1) {
+		_, _, flags := parsePlaceholder(match)
+		if flags.plus {
+			plus = true
+		}
+		if flags.query {
+			query = true
+		}
+	}
+	return
 }
 
 func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, forcePlus bool, query string, allItems []*Item) string {
@@ -1149,9 +1229,10 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, fo
 		selected = []*Item{}
 	}
 	return placeholder.ReplaceAllStringFunc(template, func(match string) string {
-		// Escaped pattern
-		if match[0] == '\\' {
-			return match[1:]
+		escaped, match, flags := parsePlaceholder(match)
+
+		if escaped {
+			return match
 		}
 
 		// Current query
@@ -1159,13 +1240,8 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, fo
 			return quoteEntry(query)
 		}
 
-		plusFlag := forcePlus
-		if match[1] == '+' {
-			match = "{" + match[2:]
-			plusFlag = true
-		}
 		items := current
-		if plusFlag {
+		if flags.plus || forcePlus {
 			items = selected
 		}
 
@@ -1201,7 +1277,9 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, fo
 					str = str[:delims[len(delims)-1][0]]
 				}
 			}
-			str = strings.TrimSpace(str)
+			if !flags.preserveSpace {
+				str = strings.TrimSpace(str)
+			}
 			replacements[idx] = quoteEntry(str)
 		}
 		return strings.Join(replacements, " ")
@@ -1257,21 +1335,30 @@ func (t *Terminal) currentItem() *Item {
 
 func (t *Terminal) buildPlusList(template string, forcePlus bool) (bool, []*Item) {
 	current := t.currentItem()
-	if !forcePlus && !hasPlusFlag(template) || len(t.selected) == 0 {
+	plus, query := hasPreviewFlags(template)
+	if !(query && len(t.input) > 0 || (forcePlus || plus) && len(t.selected) > 0) {
 		return current != nil, []*Item{current, current}
 	}
-	sels := make([]*Item, len(t.selected)+1)
-	sels[0] = current
-	for i, sel := range t.sortSelected() {
-		sels[i+1] = sel.item
+
+	// We would still want to update preview window even if there is no match if
+	//   1. command template contains {q} and the query string is not empty
+	//   2. or it contains {+} and we have more than one item already selected.
+	// To do so, we pass an empty Item instead of nil to trigger an update.
+	if current == nil {
+		current = &Item{}
+	}
+
+	var sels []*Item
+	if len(t.selected) == 0 {
+		sels = []*Item{current, current}
+	} else {
+		sels = make([]*Item, len(t.selected)+1)
+		sels[0] = current
+		for i, sel := range t.sortSelected() {
+			sels[i+1] = sel.item
+		}
 	}
 	return true, sels
-}
-
-func (t *Terminal) truncateQuery() {
-	maxPatternLength := util.Max(1, t.window.Width()-t.promptLen-1)
-	t.input, _ = t.trimRight(t.input, maxPatternLength)
-	t.cx = util.Constrain(t.cx, 0, len(t.input))
 }
 
 func (t *Terminal) selectItem(item *Item) {
@@ -1326,10 +1413,10 @@ func (t *Terminal) Loop() {
 		t.initFunc()
 		t.resizeWindows()
 		t.printPrompt()
-		t.placeCursor()
-		t.refresh()
 		t.printInfo()
 		t.printHeader()
+		t.placeCursor()
+		t.refresh()
 		t.mutex.Unlock()
 		go func() {
 			timer := time.NewTimer(t.initDelay)
@@ -1463,9 +1550,9 @@ func (t *Terminal) Loop() {
 					}
 				}
 				t.placeCursor()
+				t.refresh()
 				t.mutex.Unlock()
 			})
-			t.refresh()
 		}
 	}()
 
@@ -1628,12 +1715,12 @@ func (t *Terminal) Loop() {
 					req(reqList, reqInfo)
 				}
 			case actToggleIn:
-				if t.reverse {
+				if t.layout != layoutDefault {
 					return doAction(action{t: actToggleUp}, mapkey)
 				}
 				return doAction(action{t: actToggleDown}, mapkey)
 			case actToggleOut:
-				if t.reverse {
+				if t.layout != layoutDefault {
 					return doAction(action{t: actToggleDown}, mapkey)
 				}
 				return doAction(action{t: actToggleUp}, mapkey)
@@ -1761,12 +1848,20 @@ func (t *Terminal) Loop() {
 					mx -= t.window.Left()
 					my -= t.window.Top()
 					mx = util.Constrain(mx-t.promptLen, 0, len(t.input))
-					if !t.reverse {
-						my = t.window.Height() - my - 1
-					}
 					min := 2 + len(t.header)
 					if t.inlineInfo {
 						min--
+					}
+					h := t.window.Height()
+					switch t.layout {
+					case layoutDefault:
+						my = h - my - 1
+					case layoutReverseList:
+						if my < h-min {
+							my += min
+						} else {
+							my = h - my - 1
+						}
 					}
 					if me.Double {
 						// Double-click
@@ -1778,7 +1873,7 @@ func (t *Terminal) Loop() {
 					} else if me.Down {
 						if my == 0 && mx >= 0 {
 							// Prompt
-							t.cx = mx
+							t.cx = mx + t.xoffset
 						} else if my >= min {
 							// List
 							if t.vset(t.offset+my-min) && t.multi && me.Mod {
@@ -1830,6 +1925,12 @@ func (t *Terminal) Loop() {
 		t.mutex.Unlock() // Must be unlocked before touching reqBox
 
 		if changed {
+			if t.isPreviewEnabled() {
+				_, q := hasPreviewFlags(t.preview.command)
+				if q {
+					t.version++
+				}
+			}
 			t.eventBox.Set(EvtSearchNew, t.sort)
 		}
 		for _, event := range events {
@@ -1854,7 +1955,7 @@ func (t *Terminal) constrain() {
 }
 
 func (t *Terminal) vmove(o int, allowCycle bool) {
-	if t.reverse {
+	if t.layout != layoutDefault {
 		o *= -1
 	}
 	dest := t.cy + o